Changes after review

+++ testing/fixednotegraphscalertest.cc
@@ -7,7 +7,7 @@
 #include "common.hh"
 
 namespace {
-	Note make(int note, std::string const& text, double begin = 0, double end = 1, Note::Type type = Note::Type::NORMAL) {
+	Note make(float note, std::string const& text, double begin = 0., double end = 1., Note::Type type = Note::Type::NORMAL) {
 		auto result = Note();
 
 		result.note = note;
@@ -22,9 +22,9 @@ namespace {
 	TEST(UnitTest_FixedNoteGraphScaler, calculate_1_time_0) {
 		auto vocal = VocalTrack("Songname");
 
-		vocal.noteMin = 0;
-		vocal.noteMax = 0;
-		vocal.notes.emplace_back(make(0, "A"));
+		vocal.noteMin = 0.f;
+		vocal.noteMax = 0.f;
+		vocal.notes.emplace_back(make(0.f, "A"));
 
 		auto scaler = FixedNoteGraphScaler();
 
@@ -32,18 +32,18 @@ namespace {
 
 		const auto dimension = scaler.calculate(vocal, vocal.notes.begin(), 0);
 
-		EXPECT_EQ(0, dimension.min1);
-		EXPECT_EQ(0, dimension.max1);
-		EXPECT_EQ(0, dimension.min2);
-		EXPECT_EQ(0, dimension.max2);
+		EXPECT_EQ(0.f, dimension.min1);
+		EXPECT_EQ(0.f, dimension.max1);
+		EXPECT_EQ(0.f, dimension.min2);
+		EXPECT_EQ(0.f, dimension.max2);
 	}
 
 	TEST(UnitTest_FixedNoteGraphScaler, calculate_1_time_1) {
 		auto vocal = VocalTrack("Songname");
 
-		vocal.noteMin = 0;
-		vocal.noteMax = 0;
-		vocal.notes.emplace_back(make(0, "A"));
+		vocal.noteMin = 0.f;
+		vocal.noteMax = 0.f;
+		vocal.notes.emplace_back(make(0.f, "A"));
 
 		auto scaler = FixedNoteGraphScaler();
 
@@ -51,20 +51,20 @@ namespace {
 
 		const auto dimension = scaler.calculate(vocal, vocal.notes.begin(), 1);
 
-		EXPECT_EQ(0, dimension.min1);
-		EXPECT_EQ(0, dimension.max1);
-		EXPECT_EQ(0, dimension.min2);
-		EXPECT_EQ(0, dimension.max2);
+		EXPECT_EQ(0.f, dimension.min1);
+		EXPECT_EQ(0.f, dimension.max1);
+		EXPECT_EQ(0.f, dimension.min2);
+		EXPECT_EQ(0.f, dimension.max2);
 	}
 
 	TEST(UnitTest_FixedNoteGraphScaler, calculate_3_time_0) {
 		auto vocal = VocalTrack("Songname");
 
-		vocal.noteMin = 0;
-		vocal.noteMax = 0;
-		vocal.notes.emplace_back(make(0, "A"));
-		vocal.notes.emplace_back(make(1, "A", 1, 2));
-		vocal.notes.emplace_back(make(3, "A", 2, 3));
+		vocal.noteMin = 0.f;
+		vocal.noteMax = 0.f;
+		vocal.notes.emplace_back(make(0.f, "A"));
+		vocal.notes.emplace_back(make(1.f, "A", 1., 2.));
+		vocal.notes.emplace_back(make(3.f, "A", 2., 3.));
 
 		auto scaler = FixedNoteGraphScaler();
 
@@ -72,20 +72,20 @@ namespace {
 
 		const auto dimension = scaler.calculate(vocal, vocal.notes.begin(), 0);
 
-		EXPECT_EQ(0, dimension.min1);
-		EXPECT_EQ(3, dimension.max1);
-		EXPECT_EQ(0, dimension.min2);
-		EXPECT_EQ(3, dimension.max2);
+		EXPECT_EQ(0.f, dimension.min1);
+		EXPECT_EQ(3.f, dimension.max1);
+		EXPECT_EQ(0.f, dimension.min2);
+		EXPECT_EQ(3.f, dimension.max2);
 	}
 
 	TEST(UnitTest_FixedNoteGraphScaler, calculate_3_time_1) {
 		auto vocal = VocalTrack("Songname");
 
-		vocal.noteMin = 0;
-		vocal.noteMax = 0;
-		vocal.notes.emplace_back(make(0, "A"));
-		vocal.notes.emplace_back(make(1, "A", 1, 2));
-		vocal.notes.emplace_back(make(3, "A", 2, 3));
+		vocal.noteMin = 0.f;
+		vocal.noteMax = 0.f;
+		vocal.notes.emplace_back(make(0.f, "A"));
+		vocal.notes.emplace_back(make(1.f, "A", 1., 2.));
+		vocal.notes.emplace_back(make(3.f, "A", 2., 3.));
 
 		auto scaler = FixedNoteGraphScaler();
 
@@ -93,20 +93,20 @@ namespace {
 
 		const auto dimension = scaler.calculate(vocal, vocal.notes.begin(), 1);
 
-		EXPECT_EQ(0, dimension.min1);
-		EXPECT_EQ(3, dimension.max1);
-		EXPECT_EQ(0, dimension.min2);
-		EXPECT_EQ(3, dimension.max2);
+		EXPECT_EQ(0.f, dimension.min1);
+		EXPECT_EQ(3.f, dimension.max1);
+		EXPECT_EQ(0.f, dimension.min2);
+		EXPECT_EQ(3.f, dimension.max2);
 	}
 
 	TEST(UnitTest_FixedNoteGraphScaler, calculate_vocal_min) {
 		auto vocal = VocalTrack("Songname");
 
-		vocal.noteMin = 0;
-		vocal.noteMax = 2;
-		vocal.notes.emplace_back(make(1, "A"));
-		vocal.notes.emplace_back(make(2, "A", 1, 2));
-		vocal.notes.emplace_back(make(3, "A", 2, 3));
+		vocal.noteMin = 0.f;
+		vocal.noteMax = 2.f;
+		vocal.notes.emplace_back(make(1.f, "A"));
+		vocal.notes.emplace_back(make(2.f, "A", 1., 2.));
+		vocal.notes.emplace_back(make(3.f, "A", 2., 3.));
 
 		auto scaler = FixedNoteGraphScaler();
 
@@ -114,20 +114,20 @@ namespace {
 
 		const auto dimension = scaler.calculate(vocal, vocal.notes.begin(), 1);
 
-		EXPECT_EQ(1, dimension.min1);
-		EXPECT_EQ(3, dimension.max1);
-		EXPECT_EQ(1, dimension.min2);
-		EXPECT_EQ(3, dimension.max2);
+		EXPECT_EQ(1.f, dimension.min1);
+		EXPECT_EQ(3.f, dimension.max1);
+		EXPECT_EQ(1.f, dimension.min2);
+		EXPECT_EQ(3.f, dimension.max2);
 	}
 
 	TEST(UnitTest_FixedNoteGraphScaler, calculate_vocal_max) {
 		auto vocal = VocalTrack("Songname");
 
-		vocal.noteMin = 2;
-		vocal.noteMax = 5;
-		vocal.notes.emplace_back(make(0, "A"));
-		vocal.notes.emplace_back(make(2, "A", 1, 2));
-		vocal.notes.emplace_back(make(3, "A", 2, 3));
+		vocal.noteMin = 2.f;
+		vocal.noteMax = 5.f;
+		vocal.notes.emplace_back(make(0.f, "A"));
+		vocal.notes.emplace_back(make(2.f, "A", 1., 2.));
+		vocal.notes.emplace_back(make(3.f, "A", 2., 3.));
 
 		auto scaler = FixedNoteGraphScaler();
 
@@ -135,20 +135,20 @@ namespace {
 
 		const auto dimension = scaler.calculate(vocal, vocal.notes.begin(), 1);
 
-		EXPECT_EQ(0, dimension.min1);
-		EXPECT_EQ(3, dimension.max1);
-		EXPECT_EQ(0, dimension.min2);
-		EXPECT_EQ(3, dimension.max2);
+		EXPECT_EQ(0.f, dimension.min1);
+		EXPECT_EQ(3.f, dimension.max1);
+		EXPECT_EQ(0.f, dimension.min2);
+		EXPECT_EQ(3.f, dimension.max2);
 	}
 
 	TEST(UnitTest_FixedNoteGraphScaler, calculate_ignore_SLEEP) {
 		auto vocal = VocalTrack("Songname");
 
-		vocal.noteMin = 0;
-		vocal.noteMax = 3;
-		vocal.notes.emplace_back(make(0, "A", 0, 1));
-		vocal.notes.emplace_back(make(3, "A", 1, 2, Note::Type::SLEEP));
-		vocal.notes.emplace_back(make(1, "A", 2, 3));
+		vocal.noteMin = 0.f;
+		vocal.noteMax = 3.f;
+		vocal.notes.emplace_back(make(0.f, "A", 0., 1.));
+		vocal.notes.emplace_back(make(3.f, "A", 1., 2., Note::Type::SLEEP));
+		vocal.notes.emplace_back(make(1.f, "A", 2., 3.));
 
 		auto scaler = FixedNoteGraphScaler();
 
@@ -156,10 +156,10 @@ namespace {
 
 		const auto dimension = scaler.calculate(vocal, vocal.notes.begin(), 0);
 
-		EXPECT_EQ(0, dimension.min1);
-		EXPECT_EQ(1, dimension.max1);
-		EXPECT_EQ(0, dimension.min2);
-		EXPECT_EQ(1, dimension.max2);
+		EXPECT_EQ(0.f, dimension.min1);
+		EXPECT_EQ(1.f, dimension.max1);
+		EXPECT_EQ(0.f, dimension.min2);
+		EXPECT_EQ(1.f, dimension.max2);
 	}
 }
